added SAM as a dependency; Replaced mobileSAM with lightHQSAM
