+++ env.yml
@@ -26,4 +26,5 @@ dependencies:
     - opencv-python
     - timm
     - pynrrd
+    - git+https://github.com/facebookresearch/segment-anything.git
     - git+https://github.com/juglab/featureforest.git
