+++ setup.cfg
@@ -45,6 +45,7 @@ install_requires =
     ; pytorch
     ; torchvision
     timm
+    segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
 
 [options.packages.find]
 where = src
