--- src/featureforest/SAM/models/LightHQSAM/__init__.py
@@ -1,1 +1,0 @@
-from .setup_light_hqsam import setup
