+++ src/featureforest/SAM/models/MobileSAM/__init__.py
@@ -1,0 +1,1 @@
+from .setup_mobile_sam import setup_model
