--- src/featureforest/SAM/models/segment_anything/__init__.py
@@ -1,22 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-
-from .build_sam import (
-    build_sam,
-    build_sam_vit_h,
-    build_sam_vit_l,
-    build_sam_vit_b,
-    sam_model_registry,
-)
-from .build_sam_hq import (
-    build_sam_hq,
-    build_sam_hq_vit_h,
-    build_sam_hq_vit_l,
-    build_sam_hq_vit_b,
-    sam_hq_model_registry,
-)
-from .predictor import SamPredictor
-from .automatic_mask_generator import SamAutomaticMaskGenerator
