+++ src/featureforest/SAM/setup_model.py
@@ -2,18 +2,18 @@ from pathlib import Path
 
 import torch
 
-from .models import LightHQSAM
+from .models import MobileSAM
 
 
-def setup_lighthq_sam_model():
+def setup_mobile_sam_model():
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     print(f"running on {device}")
     # sam model (light hq sam)
-    model = LightHQSAM.setup().to(device)
+    model = MobileSAM.setup_model().to(device)
     # load weights
     weights = torch.load(
         Path(__file__).parent.joinpath(
-            "./models/weights/sam_hq_vit_tiny.pth"
+            "./models/weights/mobile_sam.pt"
         ),
         map_location=device
     )
